+++ lib/recipes/recipes_helper.rb
@@ -3,6 +3,7 @@
 require_relative '../imports_for_recipes'
 require 'slop'
 require_relative 'validate'
+require_relative 'post_bake'
 
 # Takes a block and silences any `puts` from that block
 def silenced