+++ spec/recipes_spec/helpers/match_helper.rb
@@ -24,7 +24,11 @@ end
 RSpec::Matchers.define(:match_expected_when_baked_with) do |cmd|
   match do |book|
     # run formatted command with logging silenced
-    system(cmd, %i[out err] => File::NULL)
+    if ENV['TRACE_ON']
+      system(cmd)
+    else
+      system(cmd, %i[out err] => File::NULL)
+    end
 
     # write output and run test
     actual_output = "#{__dir__}/../books/#{book}/actual_output.xhtml"